Complete swap to tinyfiledialogs
Replace all instances of native-dialog functions with calls to
tinyfiledialogs. Also build a module for plain message popups.

+++ src/gui/mod_list.rs
@@ -14,7 +14,7 @@ use json_comments::strip_comments;
 use json5;
 use handwritten_json;
 use if_chain::if_chain;
-use native_dialog::{FileDialog, MessageDialog, MessageType};
+use tinyfiledialogs as tfd;
 use opener;
 
 use serde_aux::prelude::*;
@@ -22,6 +22,7 @@ use serde_aux::prelude::*;
 use crate::gui::install;
 use crate::style;
 use crate::gui::SaveError;
+use crate::gui::dialog;
 
 mod headings;
 
@@ -151,8 +152,7 @@ impl ModList {
             &last_browsed
           } else {
             &root_dir
-          };
-          let diag = FileDialog::new().set_location(start_path);
+          }.to_str().expect("Convert path to string");
 
           match opt {
             InstallOptions::FromArchive => {
@@ -160,28 +160,29 @@ impl ModList {
               if cfg!(unix) {
                 filters.push("7z");
               }
-              if let Ok(paths) = diag.add_filter("Archive types", &filters).show_open_multiple_file() {
+
+              if let Some(paths) = tfd::open_file_dialog_multi("Select archives:", start_path, Some((&["*.tar", "*.zip", "*.7z", "*.rar"], "Archive types"))) {
                 if let Some(last) = paths.last() {
-                  self.last_browsed = last.parent().map(|p| p.to_path_buf());
+                  self.last_browsed = PathBuf::from(last).parent().map(|p| p.to_path_buf());
                 }
 
                 let mod_ids: Vec<String> = self.mods.iter().map(|(id, _)| id.clone()).collect();
                 return Command::batch(paths.iter().map(|path| {
-                  Command::perform(install::handle_archive(path.to_path_buf(), root_dir.clone(), false, mod_ids.clone()), ModListMessage::ModInstalled)
+                  Command::perform(install::handle_archive(PathBuf::from(path), root_dir.clone(), false, mod_ids.clone()), ModListMessage::ModInstalled)
                 }))
               }
 
               Command::none()
             },
             InstallOptions::FromFolder => {
-              match diag.show_open_single_dir() {
-                Ok(Some(source_path)) => {
-                  self.last_browsed = source_path.parent().map(|p| p.to_path_buf());
+              match tfd::select_folder_dialog("Select mod folder:", start_path) {
+                Some(source_path) => {
+                  self.last_browsed = PathBuf::from(&source_path).parent().map(|p| p.to_path_buf());
                   let mod_ids: Vec<String> = self.mods.iter().map(|(id, _)| id.clone()).collect();
-                  return Command::perform(install::handle_archive(source_path.to_path_buf(), root_dir.clone(), true, mod_ids), ModListMessage::ModInstalled)
+                  return Command::perform(install::handle_archive(PathBuf::from(source_path), root_dir.clone(), true, mod_ids), ModListMessage::ModInstalled)
                 },
-                Ok(None) => {},
-                _ => { ModList::make_alert("Experienced an error. Did not move given folder into mods directory.".to_owned()); }
+                None => {},
+                _ => { dialog::error("Experienced an error. Did not move given folder into mods directory."); }
               }
 
               Command::none()
@@ -189,7 +190,7 @@ impl ModList {
             _ => Command::none()
           }
         } else {
-          ModList::make_alert("No install directory set. Please set the Starsector install directory in Settings.".to_string());
+          dialog::error("No install directory set. Please set the Starsector install directory in Settings.");
           return Command::none();
         }
       },
@@ -222,71 +223,67 @@ impl ModList {
                   let mod_dir = root_dir.join("mods");
                   let raw_dest = mod_dir.join(_file_name);
                   then {
-                    match ModList::make_query(format!("A directory named {:?} already exists. Do you want to replace it?\nChoosing no will abort this operation.", _file_name)) {
-                      Ok(true) => {
-                        if !raw_dest.exists() || remove_dir_all(&raw_dest).is_ok() {
-                          self.mods.retain(|_, entry| entry.path != raw_dest);
+                    if dialog::query(format!("A directory named {:?} already exists. Do you want to replace it?\nChoosing no will abort this operation.", _file_name)) {
+                      if !raw_dest.exists() || remove_dir_all(&raw_dest).is_ok() {
+                        self.mods.retain(|_, entry| entry.path != raw_dest);
 
-                          let mod_ids: Vec<String> = self.mods.iter().map(|(id, _)| id.clone()).collect();
-                          Command::perform(install::handle_archive(path.to_path_buf(), root_dir.clone(), is_folder, mod_ids), ModListMessage::ModInstalled)
-                        } else {
-                          ModList::make_alert(format!("Failed to delete existing directory. Please check permissions on mod folder/{:?}", raw_dest));
-                          Command::none()
-                        }
-                      },
-                      _ => Command::none()
+                        let mod_ids: Vec<String> = self.mods.iter().map(|(id, _)| id.clone()).collect();
+                        Command::perform(install::handle_archive(path.to_path_buf(), root_dir.clone(), is_folder, mod_ids), ModListMessage::ModInstalled)
+                      } else {
+                        dialog::error(format!("Failed to delete existing directory. Please check permissions on mod folder/{:?}", raw_dest));
+                        Command::none()
+                      }
+                    } else {
+                      Command::none()
                     }
                   } else {
-                    ModList::make_alert(format!("Encountered an error. Could not install to {:?}", path));
+                    dialog::error(format!("Encountered an error. Could not install to {:?}", path));
                     Command::none()
                   }
                 }
               },
               install::InstallError::IDExists(current_path, intended_path, maybe_parent_path, id) => {
-                match ModList::make_query(format!("A mod with ID {} already exists. Do you want to replace it?\nChoosing no will abort this operation.", id)) {
-                  Ok(true) => {
-                    let mut commands: Vec<Command<ModListMessage>> = vec![];
+                if dialog::query(format!("A mod with ID {} already exists. Do you want to replace it?\nChoosing no will abort this operation.", id)) {
+                  let mut commands: Vec<Command<ModListMessage>> = vec![];
 
-                    if let Some(entry) = self.mods.get(&id).as_ref() {
-                      if !entry.path.exists() || remove_dir_all(&entry.path).is_ok() {
-                        if let Ok(_) = rename(&current_path, intended_path) {
-                          let mut success_message = format!("Successfully installed {}", id);
+                  if let Some(entry) = self.mods.get(&id).as_ref() {
+                    if !entry.path.exists() || remove_dir_all(&entry.path).is_ok() {
+                      if let Ok(_) = rename(&current_path, intended_path) {
+                        let mut success_message = format!("Successfully installed {}", id);
 
-                          if let Some(parent_temp_path) = maybe_parent_path {
-                            if parent_temp_path != current_path {
-                              if remove_dir_all(parent_temp_path).is_err() {
-                                success_message = format!("Successfully deleted old version and installed new version, however failed to clean up empty temporary folder at {}", current_path.to_string_lossy())
-                              }
+                        if let Some(parent_temp_path) = maybe_parent_path {
+                          if parent_temp_path != current_path {
+                            if remove_dir_all(parent_temp_path).is_err() {
+                              success_message = format!("Successfully deleted old version and installed new version, however failed to clean up empty temporary folder at {}", current_path.to_string_lossy())
                             }
                           }
-
-                          commands.push(self.queue_message(success_message, false));
-                        } else {
-                          ModList::make_alert(format!("Successfully deleted old version, however, failed to move new version's files - new version was unpacked to {}", current_path.to_string_lossy()));
                         }
+
+                        commands.push(self.queue_message(success_message, false));
                       } else {
-                        ModList::make_alert(format!("Encountered an error: failed to delete old mod directory.\n Both the old version and new version of this mod are now present, you should delete one of them to avoid issues.\nThe old version is installed at {} and the new version was unpacked to {}.", entry.path.to_string_lossy(), current_path.to_string_lossy()));
+                        dialog::notif(format!("Successfully deleted old version, however, failed to move new version's files - new version was unpacked to {}", current_path.to_string_lossy()));
                       }
                     } else {
-                      ModList::make_alert(format!("Encountered an error: could not get old mod's entry.\n Both the old version and new version of this mod may now be installed, you should delete one of them to avoid issues.\nThe new version is installed at {:?}.", current_path));
+                      dialog::error(format!("Encountered an error: failed to delete old mod directory.\n Both the old version and new version of this mod are now present, you should delete one of them to avoid issues.\nThe old version is installed at {} and the new version was unpacked to {}.", entry.path.to_string_lossy(), current_path.to_string_lossy()));
                     }
-
-                    commands.append(&mut self.parse_mod_folder());
-                    Command::batch(commands)
-                  },
-                  _ => {
-                    if let Some(parent_temp_path) = maybe_parent_path {
-                      if remove_dir_all(parent_temp_path).is_err() {
-                        println!("Failed to remove temporary directory.")
-                      }
-                    }
-
-                    Command::none()
+                  } else {
+                    dialog::error(format!("Encountered an error: could not get old mod's entry.\n Both the old version and new version of this mod may now be installed, you should delete one of them to avoid issues.\nThe new version is installed at {:?}.", current_path));
                   }
+
+                  commands.append(&mut self.parse_mod_folder());
+                  Command::batch(commands)
+                } else {
+                  if let Some(parent_temp_path) = maybe_parent_path {
+                    if remove_dir_all(parent_temp_path).is_err() {
+                      println!("Failed to remove temporary directory.")
+                    }
+                  }
+
+                  Command::none()
                 }
               },
               other => {
-                ModList::make_alert(format!("Encountered error: {:?}", other));
+                dialog::error(format!("Encountered error: {:?}", other));
                 Command::none()
               }
             }
@@ -336,7 +333,7 @@ impl ModList {
         Command::none()
       },
       ModListMessage::ParseModListError(_) => {
-        ModList::make_alert(format!("Failed to parse mods folder. Mod list has not been populated."));
+        dialog::error(format!("Failed to parse mods folder. Mod list has not been populated."));
 
         Command::none()
       },
@@ -425,12 +422,12 @@ impl ModList {
       ModListMessage::Timeout(id) => {
         if Some(id) == self.debounce {
           if self.succ_messages.len() > 0 {
-            ModList::make_alert(format!("{}", self.succ_messages.join("\n")));
+            dialog::notif(format!("{}", self.succ_messages.join("\n")));
             self.succ_messages.clear();
           }
 
           if self.err_messages.len() > 0 {
-            ModList::make_alert(format!("{}", self.err_messages.join("\n")));
+            dialog::error(format!("{}", self.err_messages.join("\n")));
             self.err_messages.clear();
           }
 
@@ -700,56 +697,6 @@ impl ModList {
 
       Command::perform(tokio::time::sleep(tokio::time::Duration::from_millis(50)), |_| { ModListMessage::Timeout(0) })
     }
-  }
-
-  pub fn make_alert(message: String) {
-    let mbox = move || {
-      MessageDialog::new()
-      .set_title("Alert:")
-      .set_type(MessageType::Info)
-      .set_text(&message)
-      .show_alert()
-      .map_err(|err| { err.to_string() })
-    };
-
-    // On windows we need to spawn a thread as the msg doesn't work otherwise
-    #[cfg(target_os = "windows")]
-    match std::thread::spawn(move || {
-      mbox()
-    }).join() {
-      Ok(Ok(())) => Ok(()),
-      Ok(Err(err)) => Err(err),
-      Err(err) => Err(err).map_err(|err| format!("{:?}", err))
-    }.unwrap();
-    // unwrap() because if this goes to hell there's not really much we can do about it...
-
-    #[cfg(not(target_os = "windows"))]
-    mbox();
-  }
-
-  pub fn make_query(message: String) -> Result<bool, String> {
-    let mbox = move || {
-      MessageDialog::new()
-      .set_type(MessageType::Warning)
-      .set_text(&message)
-      .show_confirm()
-      .map_err(|err| { err.to_string() })
-    };
-
-    // On windows we need to spawn a thread as the msg doesn't work otherwise
-    #[cfg(target_os = "windows")]
-    let res = match std::thread::spawn(move || {
-      mbox()
-    }).join() {
-      Ok(Ok(confirm)) => Ok(confirm),
-      Ok(Err(err)) => Err(err),
-      Err(err) => Err(err).map_err(|err| format!("{:?}", err))
-    };
-
-    #[cfg(not(target_os = "windows"))]
-    let res = mbox();
-
-    res
   }
 }
 
@@ -1148,7 +1095,7 @@ impl ModDescription {
       },
       ModDescriptionMessage::LinkClicked(url) => {
         if let Err(_) = opener::open(url) {
-          ModList::make_alert(format!("Failed to open update link. This could be due to a number of issues unfortunately.\nMake sure you have a default browser set for your operating system, otherwise there's not much that can be done."))
+          dialog::error(format!("Failed to open update link. This could be due to a number of issues unfortunately.\nMake sure you have a default browser set for your operating system, otherwise there's not much that can be done."))
         }
       }
     }
